# src/presentation/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Speedia OS API",
        "title": "OsApi",
        "termsOfService": "https://speedia.net/tos/",
        "contact": {
            "name": "Speedia Engineering",
            "url": "https://speedia.net/",
            "email": "eng+swagger@speedia.net"
        },
        "license": {
            "name": "Eclipse Public License v2.0",
            "url": "https://www.eclipse.org/legal/epl-2.0/"
        },
        "version": "0.0.1"
    },
    "host": "localhost:1618",
    "basePath": "/_/api",
    "paths": {
...
        "/v1/ssl/": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List ssl pairs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ssl"
                ],
                "summary": "ReadSslPairs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.SslPair"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new ssl pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ssl"
                ],
                "summary": "CreateSslPair",
                "parameters": [
                    {
                        "description": "All props are required.\u003cbr /\u003evirtualHosts may be string or []string. Alias is not allowed.\u003cbr /\u003ecertificate is a string field, i.e. ignore the structure shown.\u003cbr /\u003ecertificate and key must be in Base64.",
                        "name": "createSslPairDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSslPair"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "SslPairCreated",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/v1/ssl/vhost/": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete vhosts from a ssl pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ssl"
                ],
                "summary": "DeleteSslPairVhosts",
                "parameters": [
                    {
                        "description": "All props are required.",
                        "name": "deleteSslPairVhostsDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteSslPairVhosts"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SslPairVhostsRemoved",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/v1/ssl/{sslPairId}/": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a ssl pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ssl"
                ],
                "summary": "DeleteSslPair",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SSL Pair ID that will be deleted.",
                        "name": "sslPairId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SslPairDeleted",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/v1/vhosts/": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List virtual hosts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vhosts"
                ],
                "summary": "GetVirtualHosts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.VirtualHost"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new vhost.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vhosts"
                ],
                "summary": "CreateVirtualHost",
                "parameters": [
                    {
                        "description": "Only hostname is required.\u003cbr /\u003etype may be 'top-level', 'subdomain', 'wildcard' or 'alias'. If is not provided, it will be 'top-level'. If type is 'alias', parentHostname it will be required.",
                        "name": "createVirtualHostDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateVirtualHost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "VirtualHostCreated",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/v1/vhosts/{hostname}/": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a vhost.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vhosts"
                ],
                "summary": "DeleteVirtualHost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hostname that will be deleted.",
                        "name": "hostname",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "VirtualHostDeleted",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateSslPair": {
            "type": "object",
            "properties": {
                "certificate": {
                    "$ref": "#/definitions/entity.SslCertificate"
                },
                "key": {
                    "type": "string"
                },
                "virtualHosts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.CreateVirtualHost": {
            "type": "object",
            "properties": {
                "hostname": {
                    "type": "string"
                },
                "parentHostname": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.DeleteSslPairVhosts": {
            "type": "object",
            "properties": {
                "sslPairId": {
                    "type": "string"
                },
                "virtualHosts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "entity.SslCertificate": {
            "type": "object",
            "properties": {
                "altNames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "certificateAuthority": {
                    "type": "string"
                },
                "certificateContent": {
                    "type": "string"
                },
                "commonName": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "integer"
                },
                "issuedAt": {
                    "type": "integer"
                },
                "sslId": {
                    "type": "string"
                }
            }
        },
        "entity.SslPair": {
            "type": "object",
            "properties": {
                "certificate": {
                    "$ref": "#/definitions/entity.SslCertificate"
                },
                "chainCertificates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.SslCertificate"
                    }
                },
                "key": {
                    "type": "string"
                },
                "sslPairId": {
                    "type": "string"
                },
                "virtualHosts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "entity.VirtualHost": {
            "type": "object",
            "properties": {
                "hostname": {
                    "type": "string"
                },
                "parentHostname": {
                    "type": "string"
                },
                "rootDirectory": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" + JWT token or API key.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}